requirements.txt
about-time==4.2.1
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
alive-progress==3.1.5
appdirs==1.4.4
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.5.7
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cmake==3.29.6
comm==0.1.2
contourpy==1.2.0
cycler==0.12.1
cytoolz==0.12.2
dask==2023.11.0
dataclasses==0.6
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.13.1
flow-vis==0.1
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grapheme==0.6.0
grpcio==1.60.0
h5py==3.7.0
huggingface-hub==0.19.4
idna==3.7
imagecodecs-lite
imageio==2.31.4
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
ipykernel==6.15.0
ipython==8.10.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==5.2.0
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.3.1
lightning-utilities==0.10.0
lit==18.1.8
littleutils==0.2.2
llvmlite==0.43.0
lml==0.1.0
locket==1.0.0
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
medutils-mri==0.1.1
monai==1.3.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
numba==0.60.0
numpy==1.23.5
nvidia-ml-py==12.535.161
nvitop==1.3.2
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.10.0.84
outdated==0.2.2
packaging==23.0
pandas==1.5.2
pandas-flavor==0.6.0
parso==0.8.3
partd==1.4.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pingouin==0.5.3
platformdirs==2.5.2
plotly==5.23.0
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyexcel-io==0.6.6
pyexcel-ods==0.6.0
Pygments==2.17.2
pynndescent==0.5.13
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-ignite==0.4.13
pytorch-lightning==2.1.2
pytz==2023.3.post1
pywavelets==1.5.0
PyYAML==6.0.1
pyzmq==23.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scipy==1.11.1
seaborn==0.13.0
sentry-sdk==1.38.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.2.0
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2020.6.3
timm==0.6.12
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
triton==2.0.0
typing_extensions==4.8.0
umap-learn==0.5.6
urllib3==2.2.2
wandb==0.17.4
wcwidth==0.2.5
Werkzeug==3.0.1
xarray==2023.11.0
yarl==1.9.3
zipp==3.17.0
